third_party/blink/web_tests/http/tests/xmlhttprequest/response-document-expected.txt

Request a file as the document type with asynchronous XMLHttpRequest in a document context.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS req.status is 200
PASS req.response is non-null.
PASS result is "PASS"
PASS req.status is 200
PASS req.response is non-null.
PASS result is "PASS"
PASS req.status is 200
PASS req.response is non-null.
PASS result is "PASS"
PASS successfullyParsed is true

TEST COMPLETE
